requirements.txt
prefect==3.1.10
sentence_transformers==3.3.1
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
testcontainers==4.9.0
opensearch-py==2.8.0
datasets==3.2.0
unstructured==0.16.11
bs4==0.0.2
aiohttp==3.11.11
asyncpg==0.30.0
python-dateutil==2.9.0.post0
elasticsearch==8.17.0
scikit-learn==1.6.0
numpy==1.26.4
elastic_transport==8.15.1
torch==2.5.1
transformers==4.47.1
transformers[torch]
